pyproject.toml
[project]
name = "aorus-rag"
version = "0.1.0"
description = "AORUS MASTER 16 AI Hardware Spec Assistant — Pure Python RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-cpp-python>=0.3.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "psutil>=5.9.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
